+++ MarketInvoice.CTE/MarketInvoice.CTE.Models/Response/LoanSummary.cs
@@ -5,12 +5,12 @@ namespace MarketInvoice.CTE.Models.Response
     public class LoanSummary
     {
         [JsonProperty("weeklyRepayment")]
-        public double WeeklyRepayment { get; set; }
+        public decimal WeeklyRepayment { get; set; }
 
         [JsonProperty("totalRepaid")]
-        public double TotalRepaid { get; set; }
+        public decimal TotalRepaid { get; set; }
 
         [JsonProperty("totalInterest")]
-        public double TotalInterest { get; set; }
+        public decimal TotalInterest { get; set; }
     }
 }
